remove some components
